Introduces domain structure to algeron commands
