+++ src/main/java/org/jboss/forge/arquillian/command/algeron/provider/AlgeronProviderTestSetup.java
@@ -1,4 +1,4 @@
-package org.jboss.forge.arquillian.command.algeron;
+package org.jboss.forge.arquillian.command.algeron.provider;
 
 import org.jboss.forge.roaster.model.source.JavaClassSource;
