+++ src/main/java/org/jboss/forge/arquillian/container/model/AlgeronPublishers.java
@@ -1,9 +1,9 @@
 package org.jboss.forge.arquillian.container.model;
 
 import org.jboss.forge.addon.ui.command.UICommand;
-import org.jboss.forge.arquillian.command.algeron.AddAlgeronFolderPublisher;
-import org.jboss.forge.arquillian.command.algeron.AddAlgeronGitPublisher;
-import org.jboss.forge.arquillian.command.algeron.AddAlgeronUrlPublisher;
+import org.jboss.forge.arquillian.command.algeron.publisher.AddAlgeronFolderPublisher;
+import org.jboss.forge.arquillian.command.algeron.publisher.AddAlgeronGitPublisher;
+import org.jboss.forge.arquillian.command.algeron.publisher.AddAlgeronUrlPublisher;
 
 public enum AlgeronPublishers {
