+++ src/main/java/org/jboss/forge/arquillian/container/model/AlgeronRetriever.java
@@ -1,11 +1,11 @@
 package org.jboss.forge.arquillian.container.model;
 
 import org.jboss.forge.addon.ui.command.UICommand;
-import org.jboss.forge.arquillian.command.algeron.AddAlgeronFolderRetriever;
-import org.jboss.forge.arquillian.command.algeron.AddAlgeronGitRetriever;
-import org.jboss.forge.arquillian.command.algeron.AddAlgeronMavenRetriever;
-import org.jboss.forge.arquillian.command.algeron.AddAlgeronPactBrokerRetriever;
-import org.jboss.forge.arquillian.command.algeron.AddAlgeronUrlRetriever;
+import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronFolderRetriever;
+import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronGitRetriever;
+import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronMavenRetriever;
+import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronPactBrokerRetriever;
+import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronUrlRetriever;
 
 public enum AlgeronRetriever {
